+++ manul/src/protocol/round.rs
@@ -89,7 +89,7 @@ pub trait Protocol: 'static {
     /// An object of this type will be returned when an unattributable error happens during [`Round::finalize`].
     ///
     /// It proves that the node did its job correctly, to be adjudicated by a third party.
-    type CorrectnessProof: Send + Serialize + for<'de> Deserialize<'de> + Debug;
+    type CorrectnessProof: CorrectnessProof + Serialize + for<'de> Deserialize<'de>;
 
     /// Returns `Ok(())` if the given direct message cannot be deserialized
     /// assuming it is a direct message from the round `round_id`.
@@ -222,6 +222,18 @@ impl ProtocolError for () {
         panic!("Attempt to use an empty error type in an evidence. This is a bug in the protocol implementation.")
     }
 }
+
+/// Describes unattributable errors originating during protocol execution.
+///
+/// In the situations where no specific message can be blamed for an error,
+/// each node must generate a correctness proof proving that they performed their duties correctly,
+/// and the collection of proofs is verified by a third party.
+/// One of the proofs will necessarily be missing or invalid.
+pub trait CorrectnessProof: Debug + Clone + Send {}
+
+// A convenience implementation for protocols that don't define any errors.
+// Have to do it for `()`, since `!` is unstable.
+impl CorrectnessProof for () {}
 
 /// Message payload created in [`Round::receive_message`].
 #[derive(Debug)]